token: sort paths for reproducible extract
There is no defined order in which readdir will return the entries
of a directory. In practice, order can depend on inode number or
similar. If we run p11-kit on different files systems with similar
directory structure but different inode order the output of extract
can change.

To get a stable and reproducible output, sort the paths returned by
readdir before extracting.

Co-authored-by: Tom Dohrmann <erbse.13@gmx.de>
Signed-off-by: Paul Meyer <49727155+katexochen@users.noreply.github.com>